Fix signed/unsigned warning
This also makes Float32Data2D consistent with Float32Data3D

+++ include/astra/Float32Data2D.h
@@ -53,7 +53,7 @@ protected:
 
 	int m_iWidth;			///< width of the data (x)
 	int m_iHeight;			///< height of the data (y)
-	int m_iSize;			///< total size of the data
+	size_t m_iSize;			///< total size of the data
 
 	/** Pointer to the data block, represented as a 1-dimensional array.
 	 * Note that the data memory is "owned" by this class, meaning that the
